refactor: split install-planner (read-only) vs install-executor (mutates disk)
The previous split between install-plan.ts and install-runner.ts was
muddy: backupCollisions (writes disk) lived in install-plan, planOutputs
(pure enumeration) lived in install-runner. Re-sort by I/O concern:

- install-planner.ts: resolveComputedDefaults, detectCollisions,
  mergePrefill, missingValueKeys, defaultModuleSelections, planOutputs,
  hashValues, stableJson. Reads the temp dir / vault; never writes.

- install-executor.ts: backupCollisions, restoreBackups, runInstall,
  rollbackInstall. Every function here mutates disk.

Collision type lives with the planner (shape), BackupRecord with the
executor (only materialized by execution). Old files deleted; all
call sites in install.tsx, InstallWizard, CollisionReview, Summary,
and tests updated. Renamed tests/unit/install-plan.test.ts ->
install-planner.test.ts.

140 tests passing, typecheck clean.
