+++ source/components/CollisionReview.tsx
@@ -1,6 +1,6 @@
 import { Box, Text } from 'ink';
 import { Select, Alert } from '@inkjs/ui';
-import type { Collision } from '../core/install-plan.js';
+import type { Collision } from '../core/install-planner.js';
 
 export type CollisionAction = 'backup' | 'overwrite' | 'cancel';
