+++ source/core/install-planner.ts
@@ -1,10 +1,23 @@
+/**
+ * Install planner — pure + read-only operations.
+ *
+ * Everything here answers "what would the install do?" without touching
+ * disk beyond reading the downloaded shard's temp directory. Disk
+ * mutations (backup, write, rollback) live in `install-executor.ts`.
+ */
+
 import fsp from 'node:fs/promises';
 import path from 'node:path';
 import nunjucks from 'nunjucks';
-import type { ShardSchema, ValueDefinition, ModuleSelections } from '../runtime/types.js';
+import type {
+  ShardSchema,
+  ValueDefinition,
+  ModuleSelections,
+} from '../runtime/types.js';
 import { ShardMindError, assertNever } from '../runtime/types.js';
 import { isComputedDefault } from './schema.js';
-import { pathExists } from './fs-utils.js';
+import { resolveModules } from './modules.js';
+import { sha256 } from './fs-utils.js';
 
 export interface Collision {
   outputPath: string;
@@ -14,20 +27,11 @@ export interface Collision {
   kind: 'file' | 'directory';
 }
 
-export interface BackupRecord {
-  originalPath: string;
-  backupPath: string;
+export interface PlannedOutput {
+  outputPath: string;
+  source: 'render' | 'copy';
 }
 
-/**
- * Resolve values whose default is a computed expression (`{{ ... }}`).
- * Runs after non-computed values have been collected from the wizard or
- * a --values file.
- *
- * The expression is rendered through a standalone Nunjucks environment
- * (autoescape off) with the already-collected values as context.
- * The resulting string is coerced back into the value's declared type.
- */
 export function resolveComputedDefaults(
   schema: ShardSchema,
   collected: Record<string, unknown>,
@@ -36,7 +40,7 @@ export function resolveComputedDefaults(
   const result: Record<string, unknown> = { ...collected };
 
   for (const [key, def] of Object.entries(schema.values)) {
-    if (result[key] !== undefined) continue; // user already answered
+    if (result[key] !== undefined) continue;
     if (def.default === undefined) continue;
     if (!isComputedDefault(def.default)) continue;
 
@@ -103,8 +107,9 @@ function coerceToType(raw: string, def: ValueDefinition, key: string): unknown {
 
 /**
  * Detect which planned output paths already exist on disk.
- * Returns one Collision per existing file with size + mtime so the user
- * can make an informed choice between backup / overwrite / cancel.
+ * Flags files AND directories — a directory at a planned file path
+ * would cause EISDIR during write; the UI labels each entry so the
+ * user sees what's actually in the way.
  */
 export async function detectCollisions(
   vaultRoot: string,
@@ -116,9 +121,6 @@ export async function detectCollisions(
     const absolutePath = path.join(vaultRoot, outputPath);
     try {
       const stat = await fsp.stat(absolutePath);
-      // Flag both files and directories — a directory at a planned file
-      // path would cause EISDIR during write. The UI shows both with
-      // clear labels so the user sees the real blocker.
       if (stat.isFile() || stat.isDirectory()) {
         collisions.push({
           outputPath,
@@ -143,88 +145,6 @@ export async function detectCollisions(
   return collisions;
 }
 
-/**
- * Rename each colliding file to `<original>.shardmind-backup-<timestamp>`.
- * Timestamp is ISO-8601 compact (no colons) so the name is filesystem-safe.
- * If multiple collisions share the same timestamp suffix (same second,
- * or an earlier backup still on disk), we increment a counter to keep
- * backup names unique. Returns a record per backup so callers can
- * surface them in the summary and the rollback path can restore them.
- */
-export async function backupCollisions(
-  collisions: Collision[],
-  timestamp: Date = new Date(),
-): Promise<BackupRecord[]> {
-  const stamp = timestamp.toISOString().replace(/:/g, '-').replace(/\..+$/, '');
-  const records: BackupRecord[] = [];
-
-  for (const collision of collisions) {
-    const backupPath = await uniqueBackupPath(collision.absolutePath, stamp);
-    try {
-      await fsp.rename(collision.absolutePath, backupPath);
-    } catch (err) {
-      throw new ShardMindError(
-        `Could not back up existing file: ${collision.absolutePath}`,
-        'BACKUP_FAILED',
-        err instanceof Error ? err.message : String(err),
-      );
-    }
-    records.push({ originalPath: collision.absolutePath, backupPath });
-  }
-
-  return records;
-}
-
-async function uniqueBackupPath(absolutePath: string, stamp: string): Promise<string> {
-  const base = `${absolutePath}.shardmind-backup-${stamp}`;
-  if (!(await pathExists(base))) return base;
-  for (let i = 1; i < 1000; i++) {
-    const candidate = `${base}.${i}`;
-    if (!(await pathExists(candidate))) return candidate;
-  }
-  throw new ShardMindError(
-    `Could not find a unique backup name for ${absolutePath}`,
-    'BACKUP_FAILED',
-    'Too many existing backups with the same timestamp — clean up old .shardmind-backup-* files and retry.',
-  );
-}
-
-/**
- * Restore backups to their original paths. Used during rollback after
- * a failed install so the user's pre-install content comes back intact.
- * Best-effort per entry — individual failures are logged via the return
- * value but don't abort the rest of the restore.
- */
-export async function restoreBackups(
-  records: BackupRecord[],
-): Promise<{ restored: BackupRecord[]; failed: Array<BackupRecord & { reason: string }> }> {
-  const restored: BackupRecord[] = [];
-  const failed: Array<BackupRecord & { reason: string }> = [];
-
-  for (const record of records) {
-    try {
-      // Remove whatever the partial install wrote at the original path,
-      // then move the backup back.
-      await fsp.rm(record.originalPath, { recursive: true, force: true });
-      await fsp.rename(record.backupPath, record.originalPath);
-      restored.push(record);
-    } catch (err) {
-      failed.push({
-        ...record,
-        reason: err instanceof Error ? err.message : String(err),
-      });
-    }
-  }
-
-  return { restored, failed };
-}
-
-/**
- * Merge default values from the schema with prefill values provided via
- * `--values file.yaml`. Prefill wins over schema defaults. Computed
- * defaults are left unresolved (caller runs `resolveComputedDefaults`
- * after all prompts complete).
- */
 export function mergePrefill(
   schema: ShardSchema,
   prefill: Record<string, unknown>,
@@ -244,11 +164,6 @@ export function mergePrefill(
   return merged;
 }
 
-/**
- * Which schema value IDs still need prompting — not present in the
- * (prefill + static defaults) snapshot, not computed. Preserves schema
- * declaration order so the wizard walks values in the author's order.
- */
 export function missingValueKeys(
   schema: ShardSchema,
   snapshot: Record<string, unknown>,
@@ -262,17 +177,68 @@ export function missingValueKeys(
   return missing;
 }
 
-/**
- * Initial module selections: every module present in schema, with
- * `removable: false` locked to 'included' and `removable: true`
- * defaulting to 'included' (user unchecks to exclude).
- */
-export function defaultModuleSelections(
-  schema: ShardSchema,
-): ModuleSelections {
+export function defaultModuleSelections(schema: ShardSchema): ModuleSelections {
   const selections: ModuleSelections = {};
   for (const id of Object.keys(schema.modules)) {
     selections[id] = 'included';
   }
   return selections;
+}
+
+/**
+ * Enumerate every file that would be written given the selected modules.
+ * Reads the temp directory; does not touch the vault.
+ */
+export async function planOutputs(
+  schema: ShardSchema,
+  tempDir: string,
+  selections: ModuleSelections,
+): Promise<{
+  outputs: PlannedOutput[];
+  moduleFileCounts: Record<string, number>;
+  alwaysIncludedFileCount: number;
+}> {
+  const resolution = await resolveModules(schema, selections, tempDir);
+  const outputs: PlannedOutput[] = [];
+  const moduleFileCounts: Record<string, number> = {};
+  let alwaysIncludedFileCount = 0;
+
+  for (const id of Object.keys(schema.modules)) {
+    moduleFileCounts[id] = 0;
+  }
+
+  const tally = (entry: { outputPath: string; module: string | null }, source: 'render' | 'copy') => {
+    outputs.push({ outputPath: entry.outputPath, source });
+    if (entry.module && entry.module in moduleFileCounts) {
+      moduleFileCounts[entry.module]!++;
+    } else if (!entry.module) {
+      alwaysIncludedFileCount++;
+    }
+  };
+  for (const entry of resolution.render) tally(entry, 'render');
+  for (const entry of resolution.copy) tally(entry, 'copy');
+
+  return { outputs, moduleFileCounts, alwaysIncludedFileCount };
+}
+
+export function hashValues(values: Record<string, unknown>): string {
+  return sha256(JSON.stringify(stableJson(values)));
+}
+
+/**
+ * Recursively reorder object keys alphabetically so JSON.stringify produces
+ * a deterministic byte sequence. Arrays keep their order; primitives pass
+ * through. Unlike the `replacer` array overload of JSON.stringify, this
+ * does NOT drop nested object keys.
+ */
+function stableJson(value: unknown): unknown {
+  if (Array.isArray(value)) return value.map(stableJson);
+  if (value && typeof value === 'object') {
+    const sorted: Record<string, unknown> = {};
+    for (const key of Object.keys(value as Record<string, unknown>).sort()) {
+      sorted[key] = stableJson((value as Record<string, unknown>)[key]);
+    }
+    return sorted;
+  }
+  return value;
 }
